// samples/Dressca/dressca-frontend/src/api-client/models/post-basket-items-input-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Web API
 * Dressca の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 買い物かごにカタログアイテムを追加する処理の入力情報を表す DTO です。             
 * @export
 * @interface PostBasketItemsInputDto
 */
export interface PostBasketItemsInputDto {
    /**
     * カタログアイテム Id を取得または設定します。 1 以上の買い物かご、およびシステムに存在するカタログアイテム Id を指定してください。             
     * @type {number}
     * @memberof PostBasketItemsInputDto
     */
    'catalogItemId': number;
    /**
     * 数量を取得または設定します。 カタログアイテム Id に指定した商品が買い物かごに含まれている場合、負の値を指定すると買い物かごから指定した数だけ取り出します。 未指定の場合は 1 です。             
     * @type {number}
     * @memberof PostBasketItemsInputDto
     */
    'addedQuantity'?: number | null;
}