// samples/Dressca/dressca-frontend/src/api-client/models/put-basket-items-input-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca Web API
 * Dressca の Web API 仕様
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 買い物かごのカタログアイテム数量を変更する処理の入力情報を表す DTO です。             
 * @export
 * @interface PutBasketItemsInputDto
 */
export interface PutBasketItemsInputDto {
    /**
     * カタログアイテム Id を取得または設定します。 1 以上の買い物かご、およびシステムに存在するカタログアイテム Id を指定してください。             
     * @type {number}
     * @memberof PutBasketItemsInputDto
     */
    'catalogItemId': number;
    /**
     * 数量を取得または設定します。 0 以上の値を設定してください。             
     * @type {number}
     * @memberof PutBasketItemsInputDto
     */
    'quantity': number;
}